+++ ycc_hull/main.py
@@ -3,17 +3,21 @@ Application entry point.
 """
 import asyncio
 import os
+from typing import Any
 
 import uvicorn
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from ycc_hull.api.boats import api_boats
+from ycc_hull.api.helpers import api_helpers
 from ycc_hull.api.holidays import api_holidays
 from ycc_hull.api.members import api_members
 from ycc_hull.api.test_data import api_test_data
 from ycc_hull.config import CONFIG, LOGGING_CONFIG_FILE
 from ycc_hull.controllers.members_controller import MembersController
+from datetime import datetime
+import json
 
 app = FastAPI(
     docs_url="/docs" if CONFIG.api_docs_enabled else None,
@@ -37,6 +41,7 @@ if CONFIG.api_docs_enabled:
     }
 
 app.include_router(api_boats)
+app.include_router(api_helpers)
 app.include_router(api_holidays)
 app.include_router(api_members)